Refactoring done  - separating concerns into different files  - removed unused values
