+++ Mani-Quotify/app/src/main/java/com/mani/quotify007/view/BottomAppBar.kt
@@ -1,0 +1,40 @@
+package com.mani.quotify007.view
+
+import androidx.compose.material3.Icon
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.navigation.NavHostController
+import com.mani.quotify007.view.utils.BottomNavItem
+
+@Composable
+fun BottomAppBar(navController: NavHostController?) {
+    val items = listOf(
+        BottomNavItem.Home,
+        BottomNavItem.Search,
+        BottomNavItem.Favorites
+    )
+    androidx.compose.material3.BottomAppBar {
+        val currentRoute = navController?.currentBackStackEntry?.destination?.route
+        items.forEach { item ->
+            NavigationBarItem(
+                icon = { Icon(item.icon, contentDescription = item.title) },
+                label = { Text(item.title) },
+                selected = currentRoute == item.route,
+                onClick = {
+                    navController?.navigate(item.route) {
+                        popUpTo(navController.graph.startDestinationId)
+                        launchSingleTop = true
+                    }
+                }
+            )
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun BottomAppBarPreview() {
+    BottomAppBar(navController = null)
+}
