+++ Mani-Quotify/app/src/main/java/com/mani/quotify007/view/FavoritesScreen.kt
@@ -1,0 +1,16 @@
+package com.mani.quotify007.view
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun FavoritesScreen() {
+    Text(text = "Favorites Screen")
+}
+
+@Preview
+@Composable
+fun FavoritesScreenPreview() {
+    FavoritesScreen()
+}
