+++ Mani-Quotify/app/src/main/java/com/mani/quotify007/view/HomeScreen.kt
@@ -1,0 +1,87 @@
+package com.mani.quotify007.view
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.mani.quotify007.view.utils.StringConstants.HYPHEN_SPACE
+import com.mani.quotify007.view.utils.StringConstants.QUOTE_AUTHOR_1
+import com.mani.quotify007.view.utils.StringConstants.QUOTE_OF_THE_DAY_HEADER
+import com.mani.quotify007.view.utils.StringConstants.QUOTE_TEXT_1
+
+@Composable
+fun HomeScreen() {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp),
+        verticalArrangement = Arrangement.Center,
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Text(QUOTE_OF_THE_DAY_HEADER, fontStyle = FontStyle.Italic, fontSize = 30.sp)
+        Box(
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(250.dp)
+                .padding(top = 16.dp, start = 16.dp, end = 16.dp)
+                .clip(MaterialTheme.shapes.medium)
+                .background(
+                    brush = Brush.verticalGradient(
+                        0.0f to Color.Blue, 1.0f to Color.White, startY = 0.0f, endY = 1500.0f
+                    )
+                )
+        ) {
+            // Card content
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(16.dp)
+            ) {
+                Text(
+                    text = QUOTE_TEXT_1,
+                    style = MaterialTheme.typography.titleMedium,
+                    color = Color.White,
+                    fontFamily = FontFamily.Serif,
+                    fontWeight = FontWeight.Bold,
+                    fontSize = 18.sp,
+                    modifier = Modifier.padding(bottom = 8.dp)
+                )
+                Spacer(modifier = Modifier.height(8.dp))
+
+                Text(
+                    text = HYPHEN_SPACE+QUOTE_AUTHOR_1,
+                    fontFamily = FontFamily.Monospace,
+                    color = Color.White,
+                    fontSize = 18.sp,
+                    fontWeight = FontWeight.Normal,
+                    style = MaterialTheme.typography.labelSmall,
+                )
+            }
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun HomeScreenPreview() {
+    HomeScreen()
+}
