+++ Mani-Quotify/app/src/main/java/com/mani/quotify007/view/MainScreen.kt
@@ -1,0 +1,30 @@
+package com.mani.quotify007.view
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.navigation.compose.rememberNavController
+import com.mani.quotify007.ui.theme.QuotifyAppTheme
+
+@Composable
+fun MainScreen() {
+    val navController = rememberNavController()
+    Scaffold(
+        bottomBar = { BottomAppBar(navController) }
+    ) { innerPadding ->
+        NavigationGraph(
+            navController = navController,
+            modifier = Modifier.padding(innerPadding)
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun MainScreenPreview() {
+    QuotifyAppTheme {
+        MainScreen()
+    }
+}
