+++ Mani-Quotify/app/src/main/java/com/mani/quotify007/view/SearchScreen.kt
@@ -1,0 +1,16 @@
+package com.mani.quotify007.view
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun SearchScreen() {
+    Text(text = "Search Screen")
+}
+
+@Preview
+@Composable
+fun SearchScreenPreview() {
+    SearchScreen()
+}
