+++ Mani-Quotify/app/src/main/java/com/mani/quotify007/view/utils/StringConstants.kt
@@ -1,0 +1,8 @@
+package com.mani.quotify007.view.utils
+
+object StringConstants {
+    const val QUOTE_OF_THE_DAY_HEADER = "Quote of the Day"
+    const val HYPHEN_SPACE = "- "
+    const val QUOTE_TEXT_1 = "Success is the sum of small efforts, repeated day in and day out."
+    const val QUOTE_AUTHOR_1 = "Robert Collier"
+}
